+++ lib/loadbalancer/lib/model/health-checker-details.ts
@@ -79,6 +79,20 @@ Example: `^((?!false).|\\s)*$`
 * 
     */
   "responseBodyRegex"?: string;
+  /**
+    * Specifies if health checks should always be done using plain text instead of depending on
+* whether or not the associated backend set is using SSL.
+* <p>
+If \"true\", health checks will be done using plain text even if the associated backend set is configured
+* to use SSL.
+* <p>
+If \"false\", health checks will be done using SSL encryption if the associated backend set is configured
+* to use SSL. If the backend set is not so configured the health checks will be done using plain text.
+* <p>
+Example: `false`
+* 
+    */
+  "isForcePlainText"?: boolean;
 }
 
 export namespace HealthCheckerDetails {