+++ CHANGELOG.md
@@ -3,6 +3,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## 2.56.0 - 2023-04-04
+### Added
+- Support for pre-emptible worker nodes in the Container Engine for Kubernetes service
+- Support for larger data storage (now up to 128TB) in the MySQL Database service
+- Support for HTTP health checks for HTTPS backend sets in the Load Balancer service  
+ 
+### Breaking Changes
+- The property `backendSetName` has been changed from optional to required in the model `ForwardToBackendSet` in the Load Balancer service
+
+
 ## 2.55.1 - 2023-03-28
 ### Added
 - Support for ACD and OKV wallet naming for autonomous databases and dedicated autonomous databases on Exadata Cloud at Customer in the Database service